# api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EventManagement swagger application",
        "description": "Api d'une Application de  gestion des evenements ",
        "contact": {
            "email": "taneyemc@gmail.com"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local development server"
        }
    ],
    "paths": {
...
        "/api/user/store": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Register a new User",
                "description": "Create a new user with the provided details",
                "operationId": "1dbed59d84ed298201a3d1e168968f9a",
                "requestBody": {
                    "description": "User registration details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "email",
                                    "password",
                                    "password_confirmation",
                                    "gender",
                                    "role_id"
                                ],
                                "properties": {
                                    "firstname": {
                                        "description": "User firstname",
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "description": "User lastname",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "User email",
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "gender": {
                                        "description": "User gender",
                                        "type": "string"
                                    },
                                    "phone": {
                                        "description": "User phone",
                                        "type": "number"
                                    },
                                    "profile_image": {
                                        "description": "User profile_image",
                                        "type": "string",
                                        "format": "file"
                                    },
                                    "role_id": {
                                        "description": "User role_id",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string",
                                        "format": "password"
                                    },
                                    "password_confirmation": {
                                        "description": "User password confirmation",
                                        "type": "string",
                                        "format": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User successfully registered",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Utilisateur créé avec succès"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erreur de validation"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "firstname": [
                                                    "Le champ firstname est requis."
                                                ],
                                                "email": [
                                                    "Le champ email est requis.",
                                                    "Le champ email doit être une adresse email valide."
                                                ],
                                                "gender": [
                                                    "Le champ gender est requis."
                                                ],
                                                "password": [
                                                    "Le champ password est requis.",
                                                    "Veillez confirmer le mot de passe.",
                                                    "Le mot de passe doit avoir au moins 6 caractères."
                                                ],
                                                "role_id": [
                                                    "Le champ role est requis."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/showUser/{user}": {
            "get": {
                "tags": [
                    "user"
                ],
                "summary": "Get user by ID",
                "description": "Get user details by user ID",
                "operationId": "d3053930a078533947f82d26743486ff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/user/updateUser/{user}": {
            "put": {
                "tags": [
                    "user"
                ],
                "summary": "Update User",
                "description": "Update details of a user by their ID",
                "operationId": "ea047b27d38dca6823e2a4ead89599c4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the user",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "User data to be updated",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "email",
                                    "phone",
                                    "gender",
                                    "role_id"
                                ],
                                "properties": {
                                    "firstname": {
                                        "type": "string"
                                    },
                                    "lastname": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "email"
                                    },
                                    "phone": {
                                        "type": "number"
                                    },
                                    "gender": {
                                        "type": "string"
                                    },
                                    "role_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User successful updated."
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
        "/api/user/toggleStatus/{user}": {
            "post": {
                "tags": [
                    "user"
                ],
                "summary": "Activate/Deactivate user",
                "description": "Activate or deactivate a user account by user ID",
                "operationId": "c7577feb5f83d1d8bb295d07207e0430",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User account status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cet utilisateur a bien été activé"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "User account status updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Cet utilisateur a bien été désactivé"
                                        },
                                        "user": {
                                            "$ref": "#/components/schemas/User"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found"
                    }
                }
            }
        },
...
        "/api/auth/profile": {
            "get": {
                "tags": [
                    "auth"
                ],
                "summary": "Get user profile",
                "operationId": "profile",
                "responses": {
                    "200": {
                        "description": "User profile retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/oauth/authorize",
                        "tokenUrl": "http://localhost/oauth/token",
                        "refreshUrl": "http://localhost/token/refresh",
                        "scopes": []
                    }
                }
            },
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}